Seperated routing concerns into @AppRoutingModule class
