+++ routing-start/src/app/app.module.ts
@@ -1,26 +1,7 @@
+import { AppRoutingModule } from "./../../../course-project/src/app/app-routing.module";
 import { BrowserModule } from "@angular/platform-browser";
 import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
-import { Routes, RouterModule } from "@angular/router";
-
-const appRoutes: Routes = [
-  { path: "", component: HomeComponent },
-  {
-    path: "users",
-    component: UsersComponent,
-    children: [{ path: ":id/:name", component: UserComponent }]
-  },
-  {
-    path: "servers",
-    component: ServersComponent,
-    children: [
-      { path: ":id", component: ServerComponent },
-      { path: ":id/edit", component: EditServerComponent }
-    ]
-  },
-  { path: "not-found", component: PageNotFoundComponent },
-  { path: "**", redirectTo: "not-found" }
-];
 
 import { AppComponent } from "./app.component";
 import { HomeComponent } from "./home/home.component";
@@ -43,7 +24,7 @@ import { PageNotFoundComponent } from "./page-not-found/page-not-found.component
     ServerComponent,
     PageNotFoundComponent
   ],
-  imports: [BrowserModule, FormsModule, RouterModule.forRoot(appRoutes)],
+  imports: [BrowserModule, FormsModule, AppRoutingModule],
   providers: [ServersService],
   bootstrap: [AppComponent]
 })
